+++ apps/checklists/views.py
@@ -1,74 +1,26 @@
-import re
-
-from django.http import JsonResponse
 from django.views import generic
 
-from ebird.codes.locations import (
-    is_country_code,
-    is_location_code,
-    is_subnational1_code,
-    is_subnational2_code,
-)
+from django_filters.views import FilterView
 
-from data.models import Checklist, Country, County, Observer, State
+from data.models import Checklist, Country
+
+from .filters import ChecklistFilter
 
 
-class ChecklistsView(generic.ListView):
+class ChecklistsView(FilterView):
     model = Checklist
+    filterset_class = ChecklistFilter
     template_name = "checklists/index.html"
     paginate_by = 50
     ordering = ("-started",)
 
-    def get_filters(self):
-        filters = {
-            "country": self.request.GET.get("country"),
-            "state": self.request.GET.get("state"),
-            "county": self.request.GET.get("county"),
-            "location": self.request.GET.get("location"),
-            "observer": self.request.GET.get("observer"),
-        }
-
-        if code := self.request.GET.get("code"):
-            if is_country_code(code):
-                filters["country"] = code
-            elif is_subnational1_code(code):
-                filters["state"] = code
-            elif is_subnational2_code(code):
-                filters["county"] = code
-            elif is_location_code(code):
-                filters["location"] = code
-            elif code.startswith("USER"):
-                filters["observer"] = code
-
-        return filters
-
     def get_queryset(self):
-        qs = super().get_queryset()
-        filters = self.get_filters()
-
-        if country := filters.get("country"):
-            qs = qs.filter(country__code=country)
-        elif state := filters.get("state"):
-            qs = qs.filter(state__code=state)
-        elif county := filters.get("county"):
-            qs = qs.filter(county__code=county)
-        elif location := filters.get("location"):
-            qs = qs.filter(location__identifier=location)
-        elif observer := filters.get("observer"):
-            qs = qs.filter(observer__identifier=observer)
-
-        return qs.select_related(
-            "country",
-            "state",
-            "county",
-            "location",
-            "observer",
-        )
+        related = ["country", "state", "county", "location", "observer"]
+        queryset = super().get_queryset().select_related(*related)
+        return self.filterset_class(self.request.GET, queryset).qs
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["filters"] = self.get_filters()
-        context["search"] = self.request.GET.get("search", "")
         context["show_country"] = Country.objects.all().count() > 1
         return context
 
@@ -85,47 +37,3 @@ class DetailView(generic.DetailView):
         context = super().get_context_data(**kwargs)
         context["observations"] = context["checklist"].observations.all()
         return context
-
-
-def autocomplete(request):
-    """
-    Return the list of countries and states for the search field.
-    If there is only one country, remove it from the label.
-    """
-    data = []
-
-    for code, place in Country.objects.all().values_list("code", "place"):
-        data.append(
-            {
-                "value": code,
-                "label": place,
-            }
-        )
-
-    if len(data) == 1:
-        country = data.pop(0)["label"]
-    else:
-        country = None
-
-    for code, place in State.objects.all().values_list("code", "place"):
-        if country:
-            label = re.sub(r", %s$" % country, "", place)
-        else:
-            label = place
-        data.append({"value": code, "label": label})
-
-    for code, place in County.objects.all().values_list("code", "place"):
-        if country:
-            label = re.sub(r", %s$" % country, "", place)
-        else:
-            label = place
-        data.append({"value": code, "label": label})
-
-    observers = (
-        Observer.objects.all().values_list("identifier", "name").exclude(identifier="")
-    )
-
-    for identifier, name in observers:
-        data.append({"value": identifier, "label": name})
-
-    return JsonResponse(data, safe=False)