+++ config/settings.py
@@ -50,6 +50,8 @@ if DJANGO_ENV == "production" and DEBUG:
 # #####################
 
 INSTALLED_APPS = [
+    "dal",
+    "dal_select2",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",