+++ config/urls.py
@@ -7,6 +7,7 @@ from django.views.generic import RedirectView
 
 urlpatterns = i18n_patterns(
     path("", RedirectView.as_view(url=reverse_lazy("news:index")), name="index"),
+    path(_("autocomplete/"), include("data.urls")),
     path(_("news/"), include("news.urls")),
     path(_("checklists/"), include("checklists.urls")),
     path(_("observations/"), include("observations.urls")),